refactor(platform): extract session routing helpers
